+++ prisma/migrations/20230505070816_added_index_chain_name_type_and_created_at/migration.sql
@@ -1,0 +1,5 @@
+-- CreateIndex
+CREATE INDEX "response_codes_chain_name_type_idx" ON "response_codes"("chain_name", "type");
+
+-- CreateIndex
+CREATE INDEX "response_codes_createdAt_idx" ON "response_codes"("createdAt");