+++ prisma/migrations/20230824062617_added_new_type_singular/migration.sql
@@ -1,0 +1,2 @@
+-- AlterEnum
+ALTER TYPE "types" ADD VALUE 'singular';