+++ src/db/prisma-token.ts
@@ -1,0 +1,6 @@
+import { Token } from 'typedi';
+import { PrismaClient } from '@prisma/client';
+
+const prismaToken = new Token<PrismaClient>('Prisma');
+
+export default prismaToken;