+++ src/env-vars.ts
@@ -1,5 +1,6 @@
 import { Container, Token } from 'typedi';
-
+import fs from 'fs';
+import { chainNodeList, chainNodeList2 } from './types';
 export namespace ProcessEnvVars {
   /** Must be retrieved from the {@link Container}, and not instantiated directly. */
   export class DefaultApi {
@@ -39,5 +40,22 @@ export namespace EnvVars {
   export function getUrls(): string[] {
     const value = getEnvVar('URLS');
     return isUnset(value) ? [] : value.split(',');
+  }
+
+  export function readUrls(fileName = '/chainNodeList.json'): chainNodeList[] {
+    const configFile = fs.readFileSync(__dirname + fileName);
+    const chainNodeListJSON = JSON.parse(configFile.toString());
+    return chainNodeListJSON.chainNodeList;
+  }
+
+  export function readUrls2(fileName = '/chainNodeList.json'): chainNodeList2[] {
+    const configFile = fs.readFileSync(__dirname + fileName);
+    const chainNodeListJSON = JSON.parse(configFile.toString());
+    return chainNodeListJSON.chainNodeList;
+  }
+
+  export function getDeleteResponseDataAfterDays(): number {
+    const value = getEnvVar('DELETE_RESPONSE_DATA_AFTER_DAYS');
+    return isUnset(value) ? 30 : Number(value);
   }
 }